# docs/swagger.yaml
basePath: /api
definitions:
...
  models.DocumentResponse:
    properties:
      created_at:
        type: string
      file_format:
        type: string
      hash:
        type: string
      id:
        type: string
      original_name:
        type: string
      signed_by_team:
        $ref: '#/definitions/models.TeamShortResponse'
      signed_by_user:
        $ref: '#/definitions/models.UserShortResponse'
      updated_at:
        type: string
      verification_count:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      created_at:
        type: string
      error:
        type: string
    type: object
...
  models.TeamShortResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  models.UploadResponse:
    properties:
      file_path:
        type: string
      hash:
        type: string
      signature:
        type: string
    type: object
...
  models.UserShortResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: RESTful API for signing and verifying files with digital signatures
  title: Tawtheeq API
  version: "1.0"
paths:
...
  /documents/{id}/hide:
    get:
      consumes:
      - application/json
      description: Hide document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Hide document
      tags:
      - documents
  /documents/{id}/show:
    get:
      consumes:
      - application/json
      description: Show document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Show document
      tags:
      - documents
  /documents/hidden:
    get:
      consumes:
      - application/json
      description: Get all documents hidden
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all documents hidden
      tags:
      - documents
  /documents/me/hidden:
    get:
      consumes:
      - application/json
      description: Get all documents from me hidden
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all documents from me hidden
      tags:
      - documents
  /documents/my/{id}/hide:
    get:
      consumes:
      - application/json
      description: Hide document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Hide document
      tags:
      - documents
  /documents/myteam:
    get:
      consumes:
      - application/json
      description: Get all documents from my team
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all documents from my team
      tags:
      - documents
  /documents/myteam/{id}/hide:
    get:
      consumes:
      - application/json
      description: Hide document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Hide document
      tags:
      - documents
  /documents/team/{team_id}/hidden:
    get:
      consumes:
      - application/json
      description: Get all documents from team hidden
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all documents from team hidden
      tags:
      - documents
  /documents/team/{team_id}/visible:
    get:
      consumes:
      - application/json
      description: Get all documents from team visible
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all documents from team visible
      tags:
      - documents
  /documents/user/{user_id}/hidden:
    get:
      consumes:
      - application/json
      description: Get all documents from user hidden
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all documents from user hidden
      tags:
      - documents
  /documents/user/{user_id}/visible:
    get:
      consumes:
      - application/json
      description: Get all documents from user visible
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all documents from user visible
      tags:
      - documents
  /documents/user/me:
    get:
      consumes:
      - application/json
      description: Get all documents from me visible
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all documents from me visible
      tags:
      - documents
  /documents/visible:
    get:
      consumes:
      - application/json
      description: Get all documents visible
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all documents visible
      tags:
      - documents
...
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload file
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Upload file
      tags:
      - documents
...
  /verify/{id}:
    get:
      consumes:
      - application/json
      description: Verify file by ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Verify file by ID
      tags:
      - documents
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"